Make bottom form card flush with bottom of instructions panel
I can't target wizard__step:last-child or :last-of-type b/c those selectors are for elements, not classes. CSS doesn't us a way to target the last item of a class afaik. So I made the bottom of the form a footer instead of a div. That way the last div is the bottom card, which I want to target, and I can target it with last-of-type